+++ source/Examples/Resources/CMAssetIntelligencePoint/CMAssetIntelligencePoint_Present.ps1
@@ -1,0 +1,26 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to add an asset intelligence synchronization point to Configuration Manager.
+#>
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMAssetIntelligencePoint ExampleSettings
+        {
+            SiteCode              = 'Lab'
+            SiteServerName        = 'CA01.contoso.com'
+            Ensure                = 'Present'
+            Schedule              = DSC_CMAssetIntelligenceSynchronizationSchedule
+            {
+                RecurInterval = 'Days'
+                RecurCount    = '7'
+            }
+            Enable                = $True
+            EnableSynchronization = $True
+            IsSingleInstance      = 'Yes'
+        }
+    }
+}